+++ docs/adr/0013-phase-52-5-2-package-scaffolding-and-compatibility-shim-policy.md
@@ -98,6 +98,25 @@ This extends ADR-0013 beyond the original Phase 52.5.2 scaffold-only baseline; t
 - `aegisops_control_plane.action_review_write_surface:ActionReviewWriteSurface`
 - `aegisops_control_plane.actions.review.action_review_write_surface:ActionReviewWriteSurface`
 
+Phase 52.5.5 extends the verifier so moved action execution, lifecycle, policy, receipt, and reconciliation modules are checked through both legacy root imports and target `actions` package imports:
+
+- `aegisops_control_plane.action_lifecycle_write_coordinator:ActionLifecycleWriteCoordinator`
+- `aegisops_control_plane.actions.action_lifecycle_write_coordinator:ActionLifecycleWriteCoordinator`
+- `aegisops_control_plane.action_policy:evaluate_action_policy_record`
+- `aegisops_control_plane.actions.action_policy:evaluate_action_policy_record`
+- `aegisops_control_plane.action_receipt_validation:MissingReceiptValueError`
+- `aegisops_control_plane.actions.action_receipt_validation:MissingReceiptValueError`
+- `aegisops_control_plane.action_reconciliation_orchestration:ActionOrchestrationBoundary`
+- `aegisops_control_plane.actions.action_reconciliation_orchestration:ActionOrchestrationBoundary`
+- `aegisops_control_plane.execution_coordinator:ExecutionCoordinator`
+- `aegisops_control_plane.actions.execution_coordinator:ExecutionCoordinator`
+- `aegisops_control_plane.execution_coordinator_action_requests:ReviewedActionRequestCoordinator`
+- `aegisops_control_plane.actions.execution_coordinator_action_requests:ReviewedActionRequestCoordinator`
+- `aegisops_control_plane.execution_coordinator_delegation:ApprovedActionDelegationCoordinator`
+- `aegisops_control_plane.actions.execution_coordinator_delegation:ApprovedActionDelegationCoordinator`
+- `aegisops_control_plane.execution_coordinator_reconciliation:ActionExecutionReconciliationCoordinator`
+- `aegisops_control_plane.actions.execution_coordinator_reconciliation:ActionExecutionReconciliationCoordinator`
+
 ## 5. Layout Guardrail Skeleton
 
 The layout guardrail skeleton rejects new flat root-level Python modules unless the Phase 52.5.1 inventory or the Phase 52.5.2 approved scaffold package set classifies them.
